feat: use slice parttern for seperation of zustand states in library
